Refactor: FilterMediaByRatingUseCase
- Change `rating` parameter type from `Float` to `Double`.
- Add unit tests for `FilterMediaByRatingUseCase`.
- Update `createMedia` test utility to accept `rating` as a parameter.

+++ domain/search/src/main/java/com/domain/search/useCases/FilterMediaByRatingUseCase.kt
@@ -3,9 +3,7 @@ package com.domain.search.useCases
 import com.domain.search.model.Media
 
 class FilterMediaByRatingUseCase {
-    operator fun invoke(rating: Float, mediaList: List<Media>): List<Media> {
-        return mediaList.filter { media ->
-            media.rating >= rating
-        }
+    operator fun invoke(rating: Double, mediaList: List<Media>): List<Media> {
+        return mediaList.filter { media -> media.rating >= rating }
     }
 }
